Use setup/teardown of the folder where test is defined
When used in conjunction with --recursive, this patch makes it so
that bats take into consideration if there's a setup/teardown defined
in the files' folder.

For example, one might want to run the tests using --recursive from the
root of a dir structure like this
.
|-- graphics
|   |-- graphics.bats
|   `-- setup_suite.bash
|-- qt5
|   |-- qt5.bats
|   `-- setup_suite.bash

Which would try to execute the setup_suite of the root instead of
each directory.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>